Add one more missed rename
