+++ packages/in_app_purchase/in_app_purchase_storekit/test/in_app_purchase_storekit_platform_addtion_test.dart
@@ -7,7 +7,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:in_app_purchase_storekit/in_app_purchase_storekit.dart';
 import 'package:in_app_purchase_platform_interface/in_app_purchase_platform_interface.dart';
 
-import 'fakes/fake_ios_platform.dart';
+import 'fakes/fake_storekit_platform.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
